requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
decorator==5.0.9
discord==1.0.1
discord.py==1.7.2
fuzzywuzzy==0.18.0
idna==3.2
multidict==5.1.0
numpy==1.20.3
opencv-python==4.5.2.54
pandas==1.2.4
Pillow==8.3.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-Levenshtein==0.12.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
validators==0.18.2
webdriver-manager==3.4.2
yarl==1.6.3
